Déplacement du fichier dans la cog
